refactor: Order 도메인 패키지 구성에 따른 리펙토링 진행
