+++ src/main/java/kitchenpos/infra/DefaultPurgomalumClient.java
@@ -18,9 +18,9 @@ public class DefaultPurgomalumClient implements PurgomalumClient {
     @Override
     public boolean containsProfanity(final String text) {
         final URI url = UriComponentsBuilder.fromUriString("https://www.purgomalum.com/service/containsprofanity")
-            .queryParam("text", text)
-            .build()
-            .toUri();
+                .queryParam("text", text)
+                .build()
+                .toUri();
         return Boolean.parseBoolean(restTemplate.getForObject(url, String.class));
     }
 }
