+++ src/main/java/kitchenpos/menu/ui/MenuRestController.java
@@ -3,13 +3,7 @@ package kitchenpos.menu.ui;
 import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.domain.Menu;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
 import java.util.List;
@@ -28,7 +22,7 @@ public class MenuRestController {
     public ResponseEntity<Menu> create(@RequestBody final Menu request) {
         final Menu response = menuService.create(request);
         return ResponseEntity.created(URI.create("/api/menus/" + response.getId()))
-            .body(response);
+                .body(response);
     }
 
     @PutMapping("/{menuId}/price")
