+++ src/main/java/kitchenpos/order/deliveryorder/application/DeliveryService.java
@@ -1,0 +1,4 @@
+package kitchenpos.order.deliveryorder.application;
+
+public class DeliveryService {
+}
