+++ src/main/java/kitchenpos/order/domain/OrderRepository.java
@@ -1,4 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.order.domain;
+
+import kitchenpos.order.eatinorder.domain.OrderTable;
 
 import java.util.List;
 import java.util.Optional;
