+++ src/main/java/kitchenpos/order/eatinorder/application/EatinService.java
@@ -1,0 +1,4 @@
+package kitchenpos.order.eatinorder.application;
+
+public class EatinService {
+}
