+++ src/main/java/kitchenpos/order/eatinorder/domain/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.order.eatinorder.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
