+++ src/main/java/kitchenpos/order/eatinorder/infra/OrderTableRepository.java
@@ -1,4 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.order.eatinorder.infra;
+
+import kitchenpos.order.eatinorder.domain.OrderTable;
 
 import java.util.List;
 import java.util.Optional;
