+++ src/main/java/kitchenpos/order/infra/JpaOrderRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.order.infra;
 
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
