+++ src/main/java/kitchenpos/order/takeoutorder/application/TakeoutService.java
@@ -1,0 +1,4 @@
+package kitchenpos.order.takeoutorder.application;
+
+public class TakeoutService {
+}
