+++ src/main/java/kitchenpos/order/ui/OrderRestController.java
@@ -1,15 +1,9 @@
-package kitchenpos.ui;
+package kitchenpos.order.ui;
 
 import kitchenpos.application.OrderService;
-import kitchenpos.domain.Order;
+import kitchenpos.order.domain.Order;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
 import java.util.List;
@@ -28,7 +22,7 @@ public class OrderRestController {
     public ResponseEntity<Order> create(@RequestBody final Order request) {
         final Order response = orderService.create(request);
         return ResponseEntity.created(URI.create("/api/orders/" + response.getId()))
-            .body(response);
+                .body(response);
     }
 
     @PutMapping("/{orderId}/accept")
